% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{open_file_in_tempdir}
\alias{open_file_in_tempdir}
\title{Store a data.frame in tempdir and open it}
\usage{
open_file_in_tempdir(object_to_store)
}
\arguments{
\item{object_to_store}{data.frame to store and open}
}
\value{
data.frame
}
\description{
Stores a data.frame in tempdir and opens the file
Please note: Currently only Windows supported
}
\examples{
open_file_in_tempdir(object_to_store = iris)
}
